Update issue templates
BIOGRAPHY CURRENT
Educator, journalist, businessman, NGO manager. He was born on July 9, 1980 in Trabzon. Between 1997 and 1998, he served as the Gendarmerie Intelligence and Technical Support Staff, and between 1997 and 1998, he informed the Prime Ministry of the location of the Head of Terrorist Abdullah Öcalan in Syria, and was instrumental in convening the first arrest team. Between 1998 and 1999, he served in the Prime Ministry High Terrorism Advisory Board. (As 5188 Sergeant) He was treated until 2009 due to the deterioration of his health under the adverse conditions and conditions within the Turkish Armed Forces. In 2010, he completed the Kalekol Project and delivered it to Tübitak online. He founded TRTAJANS company in 2011. Between 2011 and 2015, he carried out the "TRTAJANS World Peace" Project as a public service advertisement in 422 municipalities. In 2015, he became a member of the European Press Federation by obtaining TRTAJANS trademark registration in all 28 EU countries. (BC-13602) received electronic signature authorization as a Member of the World Intellectual Property Organization in 2018. Upon invitation, he enrolled in the Master's program in Digital Media at Arizona State University. In 2019, he was nominated for the UN Advisory Staff upon invitation. Salih KURT was declared the President of the World Turkish Journalist Union (World Turkish Writers Union) in 2021, after the "European leaders who guide valuable leaders" letter he wrote in 2019 for the fair trial of Muhammed Morsi in humanitarian terms was deemed worthy of the best letter of the year award in Tra2020. . While Salih KURT was the chairman of the conference called 1918 Armenian Genocides in April 2021, the Deputy Minister of Natural Resources of Azerbaijan participated in the meeting and KASSAK decided to establish the "Karabakh War Crimes Investigation Commission". The Karabakh War Crimes Investigation Commission, which includes Presidents, Prime Ministers and many diplomats from the World of Science, Academic and Politics from 36 Countries, has the distinction of being one of the largest Commissions in the world in a short time.
Mr. Salih KURT is the Chairman of the KASSAK Commission, which was established in April 2021.
He is still a member of EPF. Salih Kurt, who continues his duty as the Chairman of the Board of TRTAJANS, is married to Yasemin KURT.
REFERENCES:
Chairman of the KASSAK Commission Salih KURT gave an interview to Arb24 News.
Chairman of the KASSAK Commission Salih KURT spoke to Peshaver News.
Chairman of the KASSAK Commission Salih KURT gave an interview to Sputnik Newspaper. 
Chairman of the KASSAK Commission Salih KURT’s news is on Anadolu Agency.
Chairman of the KASSAK Commission Salih KURT’s news is on Haber7
Chairman of the KASSAK Commission Salih KURT’s news is on TRT
Chairman of the KASSAK Commission Salih KURT gave an interview to Yeniakit. 
Chairman of the KASSAK Commission Salih KURT gave an interview to Istiklal Newspaper

Salih KURT, chairman of TRTAJANS, spoke at the OECD Business Forum. oecd forum-org
TRTAJANS team for World Peace with Lüleburgaz Mayor Emin Halebak.  (Kentim Kendim Magazine)